requirements.txt
aiogram
python-dotenv
requests
google-auth
google-auth-oauthlib
google-api-python-client
google-search-results
chromadb
sentence-transformers
